pyproject.toml
[project]
name = "gcs-public-share"
version = "1.0.0"
description = "Gemini CLI extension for sharing files publicly via Google Cloud Storage"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["gemini", "cli", "gcs", "google-cloud-storage", "mcp", "extension"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet",
    "Topic :: Utilities",
]
dependencies = [
    "google-cloud-storage>=2.14.0",
    "google-auth>=2.23.0",
    "fastmcp>=2.0.0",
    "mcp>=1.2.0",
]

[project.urls]
Homepage = "https://github.com/SquareShiftTech/gemini-file-share-extn-by-claude"
Repository = "https://github.com/SquareShiftTech/gemini-file-share-extn-by-claude"
Issues = "https://github.com/SquareShiftTech/gemini-file-share-extn-by-claude/issues"

[project.scripts]
gcs-share-server = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
